Removing processors module
