+++ src/app/beer_garden/router.py
@@ -16,22 +16,21 @@ The router service is responsible for:
 
 import asyncio
 import logging
-import threading
-from concurrent.futures.thread import ThreadPoolExecutor
-from functools import partial
-from typing import Dict, Union
-
 import requests
 import stomp
-from stomp.exception import ConnectFailedException
-from beer_garden.api.stomp.processors import consolidate_headers, process_send_message
+import threading
 from brewtils.models import Events, Garden, Operation, Request, System, Event
 from brewtils.schema_parser import SchemaParser
+from concurrent.futures.thread import ThreadPoolExecutor
+from functools import partial
+from stomp.exception import ConnectFailedException
+from typing import Dict, Union
 
 import beer_garden
 import beer_garden.commands
 import beer_garden.config as config
 import beer_garden.db.api as db
+import beer_garden.files
 import beer_garden.garden
 import beer_garden.local_plugins.manager
 import beer_garden.log
@@ -41,7 +40,7 @@ import beer_garden.queues
 import beer_garden.requests
 import beer_garden.scheduler
 import beer_garden.systems
-import beer_garden.files
+from beer_garden.api.stomp.transport import consolidate_headers, process_message
 from beer_garden.errors import RoutingRequestException, UnknownGardenException
 from beer_garden.events import publish
 from beer_garden.garden import get_garden, get_gardens
@@ -579,7 +578,7 @@ def _forward_stomp(operation: Operation, target_garden: Garden):
                     "client-id": target_garden.connection_params["stomp_username"]
                 },
             )
-        message, response_headers = process_send_message(operation)
+        message, response_headers = process_message(operation)
         response_headers = consolidate_headers(
             response_headers, target_garden.connection_params.get("stomp_headers")
         )
